[DDW-647] Upgrade delegation countdown rendering part
