--- source/renderer/app/components/staking/delegation-countdown/DelegationCountdownInfo.scss
@@ -1,28 +1,0 @@
-@import '../stakingConfig';
-
-.component {
-  margin: 20px;
-  text-align: center;
-}
-
-.heading {
-  @extend %accentText;
-  font-size: 19px;
-  margin-bottom: 14px;
-}
-
-.description {
-  @extend %regularText;
-  color: #5e6066;
-  margin-bottom: 24px;
-}
-
-.timeLeftDesc {
-  @extend %accentText;
-}
-
-.timeLeft {
-  @extend %accentText;
-  font-size: 36px;
-  margin-bottom: 30px;
-}
